# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Mosteben_hotel-booking-system_internal_hotel_model.Hotel:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      rooms:
        items:
          $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_room_model.Room'
        type: array
      stars:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_Mosteben_hotel-booking-system_internal_room_model.Room:
    properties:
      capacity:
        type: integer
      created_at:
        type: string
      description:
        type: string
      hotel_id:
        type: integer
      id:
        type: integer
      price_per_night:
        type: number
      room_number:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: |-
    RESTful API for a Hotel Booking System built with Go, Gin, GORM, PostgreSQL, and JWT authentication.

    Features include authentication, hotel management, room management, bookings, availability, reviews, favorites, and payments.
  title: Hotel Booking System API
  version: "1.0"
paths:
...
  /hotels:
    get:
      description: Retrieve all hotels.
      produces:
      - application/json
      responses:
        "200":
          description: Hotels retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get hotels
          schema:
            additionalProperties: true
            type: object
      summary: Get all hotels
      tags:
      - Hotels
    post:
      consumes:
      - application/json
      description: Create a new hotel.
      parameters:
      - description: Hotel data
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_hotel_model.Hotel'
      produces:
      - application/json
      responses:
        "201":
          description: Hotel created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid hotel data or validation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a hotel
      tags:
      - Hotels
...
  /hotels/{id}:
    delete:
      description: Delete an existing hotel.
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Hotel deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid hotel ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Hotel not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete hotel
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a hotel
      tags:
      - Hotels
    get:
      description: Retrieve a hotel using its ID.
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Hotel retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid hotel ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Hotel not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get hotel
          schema:
            additionalProperties: true
            type: object
      summary: Get hotel by ID
      tags:
      - Hotels
    put:
      consumes:
      - application/json
      description: Update an existing hotel.
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated hotel data
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_hotel_model.Hotel'
      produces:
      - application/json
      responses:
        "200":
          description: Hotel updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid hotel data or validation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Hotel not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a hotel
      tags:
      - Hotels
  /hotels/{id}/details:
    get:
      description: Retrieve a hotel with all of its rooms.
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Hotel details retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid hotel ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Hotel not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get hotel details
          schema:
            additionalProperties: true
            type: object
      summary: Get hotel details
      tags:
      - Hotels
...
  /hotels/search:
    get:
      description: Search hotels using hotel and room filters.
      parameters:
      - description: Hotel name
        in: query
        name: name
        type: string
      - description: City
        in: query
        name: city
        type: string
      - description: Country
        in: query
        name: country
        type: string
      - description: Hotel star rating
        in: query
        name: stars
        type: integer
      - description: Minimum room price
        in: query
        name: min_price
        type: number
      - description: Maximum room price
        in: query
        name: max_price
        type: number
      - description: Room type
        in: query
        name: room_type
        type: string
      - description: Minimum room capacity
        in: query
        name: min_capacity
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Hotel search completed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid search parameters
          schema:
            additionalProperties: true
            type: object
      summary: Search hotels
      tags:
      - Hotels
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"